[SPARK-46517][PS][TESTS][FOLLOWUPS] Reorganize `IndexingTest`: factor out `test_iloc*`
### What changes were proposed in this pull request?
factor out `test_iloc*` and add the missing parity test

### Why are the changes needed?
for test parity and parallelism

### Does this PR introduce _any_ user-facing change?
no, test-only

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #44510 from zhengruifeng/ps_test_indexing_iloc.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
